+++ src/realm/object-store/c_api/realm.cpp
@@ -129,10 +129,10 @@ RLM_API bool realm_refresh(realm_t* realm)
     });
 }
 
-RLM_API realm_t* realm_freeze(realm_t* realm)
+RLM_API realm_t* realm_freeze(const realm_t* live_realm)
 {
     return wrap_err([&]() {
-        auto& p = **realm;
+        auto& p = **live_realm;
         return new realm_t{p.freeze()};
     });
 }