+++ src/realm/object-store/list.hpp
@@ -81,8 +81,9 @@ public:
 
     Results filter(Query q) const;
 
-    // Returns a frozen copy of this result
-    List freeze(std::shared_ptr<Realm> const& realm) const;
+    // Returns a frozen copy of this List.
+    // Equivalent to producing a thread-safe reference and resolving it in the frozen realm.
+    List freeze(std::shared_ptr<Realm> const& frozen_realm) const;
 
     // Get the min/max/average/sum of the given column
     // All but sum() returns none when there are zero matching rows