+++ src/realm/object-store/results.hpp
@@ -166,8 +166,9 @@ public:
     Results snapshot() const& REQUIRES(!m_mutex);
     Results snapshot() && REQUIRES(!m_mutex);
 
-    // Returns a frozen copy of this result.
-    Results freeze(std::shared_ptr<Realm> const& realm) REQUIRES(!m_mutex);
+    // Returns a frozen copy of this result
+    // Equivalent to producing a thread-safe reference and resolving it in the frozen realm.
+    Results freeze(std::shared_ptr<Realm> const& frozen_realm) REQUIRES(!m_mutex);
 
     // Returns whether or not this Results is frozen.
     bool is_frozen() const REQUIRES(!m_mutex);
@@ -354,6 +355,9 @@ private:
 
     void evaluate_sort_and_distinct_on_collection() REQUIRES(m_mutex);
     void do_evaluate_query_if_needed(bool wants_notifications = true) REQUIRES(m_mutex);
+
+    // Shared logic between freezing and thawing Results as the Core API is the same.
+    Results import_copy_into_realm(std::shared_ptr<Realm> const& realm) REQUIRES(!m_mutex);
 
     class IteratorWrapper {
     public: